refactored building and loading of initial state
