+++ packages/opds-web-client/src/components/__tests__/OPDSCatalog-test.tsx
@@ -6,7 +6,7 @@ import { shallow } from "enzyme";
 import OPDSCatalog from "../OPDSCatalog";
 import Root, { RootProps } from "../Root";
 import buildStore from "../../store";
-import { State } from "../../reducers/index";
+import { State } from "../../state";
 import { groupedCollectionData } from "./collectionData";
 import { jsdom } from "jsdom";
 
@@ -33,40 +33,14 @@ describe("OPDSCatalog", () => {
     expect(root.props().store).toBeTruthy();
   });
 
-  it("passes store to Root if given one", () => {
+  it("passes state to Root if given one", () => {
     let store = buildStore();
+    let state = store.getState();
     let wrapper = shallow(
-      <OPDSCatalog {...props} store={store} />
+      <OPDSCatalog {...props} initialState={state} />
     );
     let root = wrapper.find<RootProps>(Root);
-    expect(root.props().store).toBe(store);
-  });
-
-  it("uses preloaded state if availabile", () => {
-    let collection = groupedCollectionData;
-    let book = groupedCollectionData.lanes[0].books[0];
-    let state: State = {
-      collection: {
-        url: collection.url,
-        data: collection,
-        isFetching: false,
-        isFetchingPage: false,
-        error: null,
-        history: []
-      },
-      book: {
-        url: book.url,
-        data: book,
-        isFetching: false,
-        error: null
-      }
-    };
-    window["__PRELOADED_STATE__"] = { catalog: state };
-    let wrapper = shallow(
-      <OPDSCatalog {...props} />
-    );
-    let root = wrapper.find<RootProps>(Root);
-    expect(root.props().store.getState()).toEqual({ catalog: state });
+    expect(root.props().store.getState()).toBe(state);
   });
 
   it("passes props to Root", () => {
